Cherry picked: Refactor Return and Sales UIMessages.(#1299)
