+++ openpos-flow/src/main/java/org/jumpmind/pos/core/ui/message/SaleUIMessage.java
@@ -13,63 +13,21 @@ import org.jumpmind.pos.core.ui.data.OrderSummary;
 
 @AssignKeyBindings
 @Data
-public class SaleUIMessage extends UIMessage {
+public class SaleUIMessage extends TransactionUIMessage {
     private static final long serialVersionUID = 1L;
-
-    private String providerKey;
-
-    private List<Total> totals;
-    private Total grandTotal;
-    private List<Total> itemCounts;
-
 
     private List<OrderSummary> orders;
     private ActionItem removeOrderAction;
 
-    private ActionItem checkoutButton;
     private ActionItem helpButton;
     private ActionItem logoutButton;
-    private ActionItem loyaltyButton;
-    private String loyaltyIDLabel;
-    private String profileIcon;
-    private List<UIMembership> memberships;
-    private boolean membershipEnabled;
-    private boolean customerMissingInfoEnabled;
-    private boolean customerMissingInfo;
-    private String customerMissingInfoIcon;
-    private String customerMissingInfoLabel;
-    private String checkMarkIcon;
-    private String noMembershipsFoundLabel;
-    private ActionItem mobileLoyaltyButton;
-    private ActionItem linkedCustomerButton;
-    private ActionItem linkedEmployeeButton;
     private ActionItem promoButton;
 
-    private boolean transactionActive = false;
-
-    private UICustomer customer;
-    private UICustomer employee;
     private AdditionalLabel taxExemptCertificateDetail;
-
-    private boolean locationEnabled;
-    private String locationOverridePrompt;
-
-    private boolean enableCollapsibleItems = true;
 
     public SaleUIMessage() {
         this.setScreenType(UIMessageType.SALE);
         this.setId("sale");
-    }
-
-    public void addTotal(String name, String amount) {
-        if (totals == null) {
-            totals = new ArrayList<>();
-        }
-        totals.add(new Total(name, amount));
-    }
-
-    public void setGrandTotal(String name, String amount) {
-        this.grandTotal = new Total(name, amount);
     }
 
     public void addOrder(OrderSummary orderSummary) {
@@ -83,14 +41,8 @@ public class SaleUIMessage extends UIMessage {
         this.taxExemptCertificateDetail = new AdditionalLabel(label, value);
     }
 
-    public void addItemCount(Total total) {
-        if (itemCounts == null) {
-            itemCounts = new ArrayList<>();
-        }
-        itemCounts.add(total);
+    public void setItemCount(String name, String amount) {
+        this.setItemCount(new Total(name, amount));
     }
-
-    public void addItemCount(String name, String amount) {
-        addItemCount(new Total(name, amount));
-    }
+    
 }
